server/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
aiosqlite==0.20.0
python-multipart==0.0.9
aiofiles==23.2.1
httpx==0.27.0          # 用于集成测试
python-dotenv==1.0.1   # 从 .env 文件加载环境变量
